https://stackoverflow.com/questions/35077463/change-temporary-directory-in-rails-4What the actual fuck
we try to change the css for this route

It fails with a weird EENOENT error ... which google says is a result of
case-insensitive filesystems.

So we look at how to change the tmp dir location, since the project
directory is in Windows, but we'd rather it stored tmp files in the
Linux container.

Jesus this is hard. And Rails should be the easy button!!!

We referenced https://stackoverflow.com/questions/35077463/change-temporary-directory-in-rails-4

but fuck if that worked. We had to put it in application.rb even though we
didn't want to. jesus why do people use this framework
it's so ... magic