Make the navigation use Bulma
This required getting it to NOT include scaffolds.scss
To do this, I commented out the include in application.css

I'm pretty confused as to why it would always include scaffolds.css and
invade all my styling with styling that it made for its scaffolds page.

Not impressed, Rails. Avdi is reading some docs.