It shows an admin button
It was supposed to be a switch, but it isn't really.
But it works I guess?

I tried installing bulma-switch with yarn and after five minutes of installing shit,
I said fuck it, what is really in here? It's a little bit of CSS
So I copied that in to the application CSS. Really, people, I don't
need a thousand things.